4-Bit microprocessor
At first, write the program in the RAM enabling CE' and PROG' only in the simulation mood. Then connect the control sequencer using the switches of the control signals. Then you can run the processor using single step or auto and see the outputs on the display. Supported instructions are : LDA, MOV, INR, AND, JM, OUT and HALT